Break bevy stuff into a separate plugin
