+++ crates/3d/Cargo.toml
@@ -15,4 +15,4 @@ bevy = {version = "0.9.1", default-features = false, features = [
   "render",
   "x11",
 ]}
-terra = {path = "../core"}
+terra = {path = "../core", features = ["bevy"]}
