+++ crates/core/Cargo.toml
@@ -8,6 +8,7 @@ repository = "https://github.com/LucasPickering/terra-rs"
 version = "0.1.0"
 
 [features]
+bevy = ["dep:bevy_ecs"]
 bin = ["dep:serde_cbor"]
 stl = ["dep:stl_io"]
 svg = ["dep:svg"]
@@ -15,6 +16,7 @@ svg = ["dep:svg"]
 [dependencies]
 anyhow = "1.0"
 assert_approx_eq = "1.1"
+bevy_ecs = {version = "0.9", optional = true, default-features = false}
 derive_more = "0.99"
 fnv = "1.0"
 indexmap = "1.6"
