+++ crates/core/src/config.rs
@@ -11,6 +11,7 @@ use validator::Validate;
 /// Configuration that defines a world gen process. Two worlds generated with
 /// same config will always be identical.
 #[derive(Copy, Clone, Debug, Serialize, Deserialize, Validate)]
+#[cfg_attr(feature = "bevy", derive(bevy_ecs::system::Resource))]
 #[serde(default)]
 pub struct WorldConfig {
     /// RNG seed to use for all randomized processes during world gen.
